+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Environment` attribute to allow environment-specific code to run during tests.
 - Added support for user agent setting in tests via the `#[UserAgent]` attribute.
 - Added `slash()` helper to factories to automatically slash content before storing.
+- Added `Mantle\Support\Registration\Post_Type_Arguments` and
+  `Mantle\Support\Registration\Taxonomy_Arguments` classes to allow for
+  registering post types/taxonomies with custom arguments fluently.
 
 ### Changed